+++ backends/simplego/exec_test.go
@@ -16,64 +16,66 @@ func TestBuilder_Compile(t *testing.T) {
 	// backend must be exclusive (not shared across tests) for this test to work.
 	backend := New("")
 	builder := backend.Builder("test")
-	var x, c backends.Op
-	require.NotPanics(t, func() {
-		x = builder.Parameter("x", shapes.Make(dtypes.Float32, 3))
-		x = builder.Neg(x)
-		c = builder.Constant([]int64{1, 2, 3}, 3)
-	})
+	x, err := builder.Parameter("x", shapes.Make(dtypes.Float32, 3))
+	require.NoError(t, err)
 	require.NotNil(t, x)
+	x, err = builder.Neg(x)
+	require.NoError(t, err)
+	require.NotNil(t, x)
+	c, err := builder.Constant([]int64{1, 2, 3}, 3)
+	require.NoError(t, err)
 	require.NotNil(t, c)
-	var exec *Executable
-	require.NotPanics(t, func() { exec = builder.Compile(x, c).(*Executable) })
+
+	exec, err := builder.Compile(x, c)
+	require.NoError(t, err)
 	require.NotNil(t, exec)
 
 	// Check that it fails if fed the wrong number of parameters.
-	require.Panics(t, func() {
-		inputs := []backends.Buffer{
-			backend.BufferFromFlatData(0, []float32{1, 2, 3}, shapes.Make(dtypes.Float32, 3)),
-			backend.BufferFromFlatData(0, []int64{1, 2, 3}, shapes.Make(dtypes.Float32, 3)),
-		}
-		exec.Execute(inputs, []bool{true, true})
-	})
+	i0, err := backend.BufferFromFlatData(0, []float32{1, 2, 3}, shapes.Make(dtypes.Float32, 3))
+	require.NoError(t, err)
+	i1, err := backend.BufferFromFlatData(0, []float32{1, 2, 3}, shapes.Make(dtypes.Float32, 3))
+	require.NoError(t, err)
+	_, err = exec.Execute([]backends.Buffer{i0, i1}, []bool{true, true})
+	require.Error(t, err)
 
 	// Check that it fails if fed incompatible parameters.
-	require.Panics(t, func() {
-		inputs := []backends.Buffer{
-			backend.BufferFromFlatData(0, []float32{1, 2, 3, 4}, shapes.Make(dtypes.Float32, 4)),
-		}
-		exec.Execute(inputs, []bool{true})
-	})
-	require.Panics(t, func() {
-		inputs := []backends.Buffer{
-			backend.BufferFromFlatData(0, []int64{1, 2, 3}, shapes.Make(dtypes.Float32, 3)),
-		}
-		exec.Execute(inputs, []bool{true})
-	})
+	i0, err = backend.BufferFromFlatData(0, []float32{1, 2, 3, 4}, shapes.Make(dtypes.Float32, 4))
+	_, err = exec.Execute([]backends.Buffer{i0}, []bool{true})
+	require.Error(t, err)
+
+	i0, err = backend.BufferFromFlatData(0, []uint32{1, 2, 3}, shapes.Make(dtypes.Uint32, 3))
+	require.NoError(t, err)
+	_, err = exec.Execute([]backends.Buffer{i0}, []bool{true})
+	require.Error(t, err)
 
 	// Checks correct execution with donated inputs.
-	inputs := []backends.Buffer{
-		backend.BufferFromFlatData(0, []float32{3, 5, 7}, shapes.Make(dtypes.Float32, 3)),
-	}
-	inputsData := inputs[0].(*Buffer).flat.([]float32)
-	outputs := exec.Execute(inputs, []bool{true})
+	i0, err = backend.BufferFromFlatData(0, []float32{3, 5, 7}, shapes.Make(dtypes.Float32, 3))
+	require.NoError(t, err)
+	i0Data := i0.(*Buffer).flat.([]float32)
+	outputs, err := exec.Execute([]backends.Buffer{i0}, []bool{true})
+	require.NoError(t, err)
 	require.Len(t, outputs, 2)
-	require.True(t, &inputsData[0] == &(outputs[0].(*Buffer).flat.([]float32))[0])
-	require.True(t, backend.BufferShape(outputs[1]).Equal(shapes.Make(dtypes.Int64, 3)))
+	require.True(t, &i0Data[0] == &(outputs[0].(*Buffer).flat.([]float32))[0])
+	outputShape, err := backend.BufferShape(outputs[1])
+	require.NoError(t, err)
+	require.True(t, outputShape.Equal(shapes.Make(dtypes.Int64, 3)))
 
 	// Save reference to buffer before finalizing it: it should be re-used on the next call.
 	oldOutput1 := outputs[1].(*Buffer)
-	backend.BufferFinalize(outputs[1]) // Return buffer: we want it re-used at the next call.
+	err = backend.BufferFinalize(outputs[1]) // Return buffer: we want it re-used at the next call.
+	require.NoError(t, err)
 
 	// Checks correct execution without donated inputs.
 	// Notice the inputs were donated in the last interation, so we have to set them again.
-	inputs = []backends.Buffer{
-		backend.BufferFromFlatData(0, []float32{3, 5, 7}, shapes.Make(dtypes.Float32, 3)),
-	}
-	outputs = exec.Execute(inputs, []bool{false})
+	i0, err = backend.BufferFromFlatData(0, []float32{3, 5, 7}, shapes.Make(dtypes.Float32, 3))
+	require.NoError(t, err)
+	outputs, err = exec.Execute([]backends.Buffer{i0}, []bool{false})
+	require.NoError(t, err)
 	require.Len(t, outputs, 2)
-	require.True(t, inputs[0].(*Buffer) != outputs[0].(*Buffer))
-	require.True(t, backend.BufferShape(outputs[1]).Equal(shapes.Make(dtypes.Int64, 3)))
+	require.True(t, i0.(*Buffer) != outputs[0].(*Buffer))
+	outputShape, err = backend.BufferShape(outputs[1])
+	require.NoError(t, err)
+	require.True(t, outputShape.Equal(shapes.Make(dtypes.Int64, 3)))
 
 	// Checks that the output buffer for output1 was reused from the pool.
 	newOutput1 := outputs[1].(*Buffer)