+++ docs/CHANGELOG.md
@@ -13,7 +13,7 @@
 * Package `types/tensors`:
   * Removed dependency to `gopjrt/pjrt` -- otherwise we'll always need to install the C/C++ library.
 * Package `backend`:
-  * `Backend` interface now returns errors instead of panic'ing
+  * `Backend` interface now returns errors instead of panicking.
 * Package `graph`:
   * Added `NewExecOrError` and `Exec.CallOrError` as error-returning alternatives.
 * gofmt cleanups by @zjtv